refactor: mv rebase cmds to dedicated module
